+++ src/br/edu/up/Exercicios/Exercicio6.java
@@ -7,10 +7,10 @@ public class Exercicio6 {
         try (Scanner Preco = new Scanner(System.in)) {
 
             System.out.println("Digite o porcentagem que deseja que seja aplicado para o custo de venda(%): ");
-            double Porcentagem = Preco.nextInt();
+            double Porcentagem = Preco.nextDouble();
 
             System.out.println("Digite o Valor de costo do produto: ");
-            double Valor = Preco.nextInt();
+            double Valor = Preco.nextDouble();
 
             double Venda = Valor / (100);
